RFC1000: Switch to confirm/reject flow, add describe algo (#18)
* RFC1000: Switch to confirm/reject flow, add describe algo

This changes how `DESCRIBE CURRENT MIGRATION` works with respect to the
system-proposed DDL.  Instead of generating multiple variants of the
next proposed statement, only one statement is generated, and the client
will have a choice to either accept or reject it by issuing `ALTER
CURRENT MIGRATION REJECT PROPOSED`.  The reason for this change is that
in the general case it is very difficult to come up with a reasonable
set of multiple alternatives of the *next* valid DDL statement, because
different operations have different dependency path, even on relatively
simple deltas.  The new approach sidesteps that problem by not
attempting to present alternatives and instead guides the client with a
series of questions.