Remove half-baked "security implications"
Each one of these points is both a weak argument, and not well
constructed. URL "validation" is really a misnomer, and not a
security-related feature. Lack of rate limiting is more of an abuse
vector than a security issue, and should really be handled at the
queuing level of abstraction. TLS is a protocol-specific concern for
those protocols that include TLS affordances.